requirements.txt
## Core Framework
fastapi
uvicorn[standard]
python-multipart

## Configuration
pydantic-settings
python-dotenv

## EML Parsing
eml-parser

## Logging
loguru

## Telemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp

## Testing
pytest
httpx

## Linting & Formatting
black
flake8